filter G_GNUC before PRIVATE so $ is still true.
Sun Nov 28 13:13:56 2004  Manish Singh  <yosh@gimp.org>

        * glib/abicheck.sh: filter G_GNUC before PRIVATE so $ is still true.

        * tests/utf8-validate.c: cast pointer math to gint for error print
        message.

+++ tests/utf8-validate.c
@@ -301,7 +301,7 @@ do_test (gint         index,
 	       index,
 	       str->str, max_len, 
 	       valid ? "TRUE" : "FALSE", offset,
-	       result ? "TRUE" : "FALSE", end - text);
+	       result ? "TRUE" : "FALSE", (gint) (end - text));
       g_string_free (str, FALSE);
     }
 }
